FIRST IMPLEMENTATION
Create a bidimensional array do hold the objects, making it possible to
keep track of the properties later.
In this array, the first index contains rows, represented as a div, and
the second item contains the squares, which had to be encapsulated in a
parent div for stretching in the cross axis.